fiber-pon-tracker-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
boto3==1.35.6
